controllers almost done

+++ src/Controllers/locationController.js
@@ -1,17 +1,26 @@
 const locationService = require('../Services/locationService')
+const {validateBody, validateId} = require('../Validators/locationValidator')
 
 const createLocation = async (req, res) => {
-    const ID = req.params.id
     const body = req.body
-     
-    try {
-       const resCreated = await locationService.createLocation(ID,body)
-        res.status(200).send({data: resCreated})
-    } catch (err) {
-        res
-        .status(err?.status || 500)
-        .send({ status: "FAILED", data: { error: err?.message || err } });
+    const {status, error} = validateBody(body)
+    if(status){
+        try {
+            const {coordinates, coordinates_delta} = body
+            const geoP = locationService.toGeoPoint(coordinates)
+            const geoP_delta = locationService.toGeoPoint(coordinates_delta)
+            const location = { ...body, coordinates: geoP, coordinates_delta: geoP_delta}
+            const resCreated = await locationService.createLocation(location)
+            res.status(200).send({message: 'Location succesfully created', data: resCreated})
+        } catch (err) {
+            res
+            .status(err?.status || 500)
+            .send({ status: "FAILED", data: { error: err?.message || err } });
+        }
+    }else{
+        res.status(400).send({status: "Failed type validation", data: error })
     }
+      
 };
 
 const getLocations = async(req,res) => {
@@ -32,19 +41,25 @@ const getLocations = async(req,res) => {
 };
 
 const getLocation = async (req,res) => {
-    const id = req.params.id
-    try {
-       const location = await locationService.getLocation(id)
-        res.status(200).send({data: location})
-    } catch (err) {
-        res
-        .status(err?.status || 500)
-        .send({ status: "FAILED", data: { error: err?.message || err } });
+    const {id} = req.params
+    const {status, error} = validateId(id)
+    if(status){
+        try {
+            const location = await locationService.getLocation(id)
+            res.status(200).send({data: location})
+        } catch (err) {
+            res
+            .status(err?.status || 500)
+            .send({ status: "FAILED", data: { error: err?.message || err } });
+        }
+    }else{
+        res.status(400).send({status: "Failed type validation", data: error })
     }
+    
 };
 
 const updateLocation = async(req,res) => {
-    const id = req.params.id
+    const {id} = req.params
     const body = req.body
     try {
        const updatedlocation = await locationService.updateLocation(id, body)
@@ -52,14 +67,18 @@ const updateLocation = async(req,res) => {
     } catch (err) {
         res
         .status(err?.status || 500)
-        .send({ status: "FAILED", data: { error: err?.message || err } });
+        .send({ status: "FAILED", data: { error: err?.message || err.error.details } });
     }
 };
 
 const deleteLocation = async (req,res) => {
-    const id = req.params.id
+    const {id} = req.params
     try {
-       const deletedlocation = await locationService.deleteLocation(id)
+        /* const {data:{exists}} = await getLocation(id)
+        if(!exists){ return 'no existe el doc a eliminar' } */
+        //Revisar con el service si existe y si es asi, eliminarlo
+        //Si no existe, mandar msj que el ide buscado no existe
+        const deletedlocation = await locationService.deleteLocation(id)
         res.status(200).send({data: deletedlocation})
     } catch (err) {
         res
